Sprint 2 second version

+++ src/main/java/ru/yandex/praktikum/model/Discountable.java
@@ -2,6 +2,6 @@ package ru.yandex.praktikum.model;
 
 public interface Discountable {
 
-    int getDiscount();
+    double getDiscount();
 
 }
